restore single-factorized hamiltonian tests
single- or double- factorization does not support complex two-body operators

+++ tests/python/hamiltonians/single_factorized_hamiltonian_test.py
@@ -33,7 +33,9 @@ def test_linear_operator(norb: int, nelec: tuple[int, int], cholesky: bool):
     """Test linear operator."""
     rng = np.random.default_rng(2474)
 
-    mol_hamiltonian = ffsim.random.random_molecular_hamiltonian(norb, seed=rng)
+    mol_hamiltonian = ffsim.random.random_molecular_hamiltonian(
+        norb, seed=rng, dtype=float
+    )
     sf_hamiltonian = ffsim.SingleFactorizedHamiltonian.from_molecular_hamiltonian(
         mol_hamiltonian, cholesky=cholesky
     )
@@ -55,7 +57,9 @@ def test_reduced_matrix_product_states(norb: int, nelec: tuple[int, int]):
     """Test computing reduced matrix on product states."""
     rng = np.random.default_rng(7869)
 
-    mol_hamiltonian = ffsim.random.random_molecular_hamiltonian(norb, seed=rng)
+    mol_hamiltonian = ffsim.random.random_molecular_hamiltonian(
+        norb, seed=rng, dtype=float
+    )
     sf_hamiltonian = ffsim.SingleFactorizedHamiltonian.from_molecular_hamiltonian(
         mol_hamiltonian
     )
@@ -98,7 +102,9 @@ def test_expectation_product_state_slater_determinant(
     """Test computing expectation value on Slater determinant product state."""
     rng = np.random.default_rng(3400)
 
-    mol_hamiltonian = ffsim.random.random_molecular_hamiltonian(norb, seed=rng)
+    mol_hamiltonian = ffsim.random.random_molecular_hamiltonian(
+        norb, seed=rng, dtype=float
+    )
     sf_hamiltonian = ffsim.SingleFactorizedHamiltonian.from_molecular_hamiltonian(
         mol_hamiltonian
     )
@@ -128,7 +134,9 @@ def test_expectation_product_state(norb: int, nelec: tuple[int, int]):
     """Test computing expectation value on product state."""
     rng = np.random.default_rng(6775)
 
-    mol_hamiltonian = ffsim.random.random_molecular_hamiltonian(norb, seed=rng)
+    mol_hamiltonian = ffsim.random.random_molecular_hamiltonian(
+        norb, seed=rng, dtype=float
+    )
     sf_hamiltonian = ffsim.SingleFactorizedHamiltonian.from_molecular_hamiltonian(
         mol_hamiltonian
     )
